factureo_backend/requirements.txt
flask>=3.0.0
flask-login>=0.6.3
pymongo[srv]>=4.6.0
python-dotenv>=1.0.1
werkzeug>=3.0.0
